float 64

+++ data.py
@@ -126,8 +126,8 @@ class DataGenerator(tensorflow.keras.utils.Sequence):
     def __data_generation(self, temp_list):
         import tensorflow as tf
         'Generates data containing batch_size samples'
-        inputs = np.zeros((self.batch_size, self.img_size[0], self.img_size[1], len(self.inputs)), dtype=np.float32)
-        outputs = np.zeros((self.batch_size, self.img_size[0], self.img_size[1], len(self.outputs)), dtype=np.float32)
+        inputs = np.zeros((self.batch_size, self.img_size[0], self.img_size[1], len(self.inputs)), dtype=np.float64)
+        outputs = np.zeros((self.batch_size, self.img_size[0], self.img_size[1], len(self.outputs)), dtype=np.float64)
 
         if (len(temp_list) != self.batch_size):
             raise ValueError("Batch size is not equal to the number of files in the list")
@@ -136,7 +136,7 @@ class DataGenerator(tensorflow.keras.utils.Sequence):
         for i, ID in enumerate(temp_list):
             with np.load(ID, allow_pickle=True) as npzfile:
                 for idx in range(len(self.inputs)):
-                    inputs[i, :, :, idx] = npzfile[self.inputs[idx]].astype(np.float32)
+                    inputs[i, :, :, idx] = npzfile[self.inputs[idx]].astype(np.float64)
 
                 for idx in range(len(self.outputs)):   
 
@@ -160,10 +160,10 @@ class DataGenerator(tensorflow.keras.utils.Sequence):
                     patients_outp[patient] = np.zeros((200, 256, 256, len(self.outputs)))
 
                 for idx in range(len(self.inputs)):
-                    patients_inp[patient][slice, :, :, idx] = npzfile[self.inputs[idx]].astype(np.float32)
+                    patients_inp[patient][slice, :, :, idx] = npzfile[self.inputs[idx]].astype(np.float64)
 
                 for idx in range(len(self.outputs)):
-                    patients_outp[patient][slice, :, :, idx] = npzfile[self.outputs[idx]].astype(np.float32)
+                    patients_outp[patient][slice, :, :, idx] = npzfile[self.outputs[idx]].astype(np.float64)
                         
                 npzfile.close()
